Compilation error fixes after synchronzing with huawei-demo branch.

+++ cocos/base/CCDirector.cpp
@@ -1322,7 +1322,7 @@ void DisplayLinkDirector::stopAnimation()
     _invalid = true;
 }
 
-void DisplayLinkDirector::setAnimationInterval(float interval)
+void DisplayLinkDirector::setAnimationInterval(double interval)
 {
     setAnimationInterval(interval, SetIntervalReason::BY_GAME);
 }
